+++ src/clingexplaid/mus/utils.py
@@ -1,7 +1,8 @@
 """Utilities for the MUS functionality of clingexplaid"""
 
-from collections.abc import Iterable
+from collections.abc import Iterable, Iterator
 from dataclasses import dataclass
+from typing import override
 
 from clingo import Symbol
 
@@ -16,6 +17,53 @@ class AssumptionWrapper:
 
     def __hash__(self) -> int:
         return self.literal
+
+
+@dataclass
+class UnsatisfiableSubset:
+    """Container class for unsatisfiable subsets"""
+
+    assumptions: set[AssumptionWrapper]
+    minimal: bool = False
+
+    @staticmethod
+    def _render_assumption(assumption: AssumptionWrapper) -> str:  # nocoverage
+        assumption_sign = "+" if assumption.sign else "-"
+        return f"{assumption.symbol}[{assumption.literal},{assumption_sign}]"
+
+    @staticmethod
+    def _render_assumption_set(
+        assumptions: set[AssumptionWrapper],
+    ) -> str:  # nocoverage
+        out = "{"
+        out += ",".join([UnsatisfiableSubset._render_assumption(a) for a in assumptions])
+        out += "}"
+        return out
+
+    def iter_symbols(self) -> Iterator[tuple[Symbol, bool]]:
+        """Iterate over all assumption symbols in the unsatisfiable subset"""
+        return ((a.symbol, a.sign) for a in self.assumptions)
+
+    def iter_literals(self) -> Iterator[tuple[int, bool]]:  # nocoverage
+        """Iterate over all assumption literals in the unsatisfiable subset"""
+        return ((a.literal, a.sign) for a in self.assumptions)
+
+    def __iter__(self) -> Iterator[tuple[Symbol, bool] | int]:
+        return self.iter_symbols()
+
+    @override
+    def __str__(self) -> str:  # nocoverage
+        out = "UnsatisfiableSubset("
+        out += "assumptions="
+        out += UnsatisfiableSubset._render_assumption_set(self.assumptions)
+        out += ", minimal="
+        out += str(self.minimal)
+        out += ")"
+        return out
+
+    @override
+    def __repr__(self) -> str:
+        return self.__str__()
 
 
 def unwrap(wrapped: Iterable[AssumptionWrapper]) -> set[int]:
